reverted to selectionBoxV1
