+++ src/phaser/selectionBox.ts
@@ -1,5 +1,23 @@
 import Phaser from "phaser";
 
+// STEP 1: Collaborative Context Merging - Basic interfaces and ownership structure
+// Interface for collaborative context data
+interface BoxContextData {
+  value: any;
+  owner: string; // box id that owns this data
+  version: number;
+  lastModified: number;
+  canShare: boolean; // whether this data can be shared with neighbors
+}
+
+// Interface for box context with ownership tracking
+interface BoxContext {
+  id: string;
+  data: Map<string, BoxContextData>;
+  chatHistory: any[];
+  version: number;
+}
+
 export class SelectionBox {
   private graphics: Phaser.GameObjects.Graphics;
   private start: Phaser.Math.Vector2;
@@ -14,12 +32,17 @@ export class SelectionBox {
   private zLevel: number;
   public selectedTiles: number[][] = [];
   private layer: Phaser.Tilemaps.TilemapLayer;
-  public localContext: { chatHistory: any[] };
+  public localContext: BoxContext;
   public placedTiles: {
     tileIndex: number;
     x: number;
     y: number;
     layerName: string;
+  }[] = [];
+  public placedEnemies: {
+    enemyType: string;
+    x: number;
+    y: number;
   }[] = [];
   private tabContainer: Phaser.GameObjects.Container | null = null;
   private onSelect?: (box: SelectionBox) => void;
@@ -28,15 +51,30 @@ export class SelectionBox {
   private isActive: boolean = false;
   private isFinalized: boolean = false;
 
+  // STEP 3: Collaborative Context Merging - Neighbor tracking
+  private neighbors: Set<SelectionBox> = new Set();
+  // Intersections with boxes on different z-levels
+  private intersections: Set<SelectionBox> = new Set();
+  private lastNeighborCheck: number = 0;
+  private neighborCheckInterval: number = 500; // Check every 500ms
+
   // Drag helpers
   private _dragInitialStart?: Phaser.Math.Vector2;
   private _dragInitialEnd?: Phaser.Math.Vector2;
-  private _dragInitialContainerX?: number;
-  private _dragInitialContainerY?: number;
   private _dragPointerTileX?: number;
   private _dragPointerTileY?: number;
-  private _dragPointerOffsetX?: number;
-  private _dragPointerOffsetY?: number;
+  // Snapshot of placedTiles at the start of a drag so we can keep them in sync
+  private _dragOriginalPlacedTiles?: {
+    tileIndex: number;
+    x: number;
+    y: number;
+    layerName: string;
+  }[];
+  private _dragOriginalPlacedEnemies?: {
+    enemyType: string;
+    x: number;
+    y: number;
+  }[];
   private _dragStartHandler?: (pointer: Phaser.Input.Pointer, obj: any) => void;
   private _dragHandler?: (
     pointer: Phaser.Input.Pointer,
@@ -75,10 +113,274 @@ export class SelectionBox {
     this.redraw();
 
     // Initialize localContext with its own chatHistory
-    this.localContext = { chatHistory: [] };
+    this.localContext = {
+      id: `box_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`,
+      data: new Map(),
+      chatHistory: [],
+      version: 1,
+    };
     this.onSelect = onSelect;
     // create tab after initial draw
     this.createTab();
+  }
+
+  // STEP 2: Collaborative Context Merging - Basic data management methods
+
+  /**
+   * Set data in this box's context with ownership tracking
+   */
+  public setContextData(
+    key: string,
+    value: any,
+    canShare: boolean = false,
+  ): void {
+    const contextData: BoxContextData = {
+      value,
+      owner: this.localContext.id,
+      version: 1,
+      lastModified: Date.now(),
+      canShare,
+    };
+
+    this.localContext.data.set(key, contextData);
+    this.localContext.version++;
+  }
+
+  /**
+   * Get data from this box's context
+   */
+  public getContextData(key: string): any {
+    const contextData = this.localContext.data.get(key);
+    return contextData ? contextData.value : null;
+  }
+
+  /**
+   * Check if this box owns a specific piece of data
+   */
+  public ownsData(key: string): boolean {
+    const contextData = this.localContext.data.get(key);
+    return contextData ? contextData.owner === this.localContext.id : false;
+  }
+
+  /**
+   * Get all shareable data from this box
+   */
+  public getShareableData(): Map<string, BoxContextData> {
+    const shareableData = new Map<string, BoxContextData>();
+    this.localContext.data.forEach((data, key) => {
+      if (data.canShare) {
+        shareableData.set(key, data);
+      }
+    });
+    return shareableData;
+  }
+
+  // STEP 3: Collaborative Context Merging - Neighbor detection and management
+
+  /**
+   * Check if this box is touching another box (adjacent or overlapping)
+   */
+  public isTouching(other: SelectionBox): boolean {
+    if (other === this || other.getZLevel() !== this.zLevel) {
+      return false;
+    }
+
+    const myBounds = this.getBounds();
+    const otherBounds = other.getBounds();
+
+    // Check if boxes are adjacent (touching edges) or overlapping
+    const touching =
+      // Overlapping
+      Phaser.Geom.Intersects.RectangleToRectangle(myBounds, otherBounds) ||
+      // Adjacent horizontally
+      ((myBounds.right + 1 === otherBounds.left ||
+        otherBounds.right + 1 === myBounds.left) &&
+        !(
+          myBounds.bottom < otherBounds.top || otherBounds.bottom < myBounds.top
+        )) ||
+      // Adjacent vertically
+      ((myBounds.bottom + 1 === otherBounds.top ||
+        otherBounds.bottom + 1 === myBounds.top) &&
+        !(
+          myBounds.right < otherBounds.left || otherBounds.right < myBounds.left
+        ));
+
+    return touching;
+  }
+
+  /**
+   * Update neighbor list by checking all boxes in the scene
+   */
+  public updateNeighbors(allBoxes: SelectionBox[]): void {
+    const now = Date.now();
+    if (now - this.lastNeighborCheck < this.neighborCheckInterval) {
+      return; // Don't check too frequently
+    }
+
+    this.lastNeighborCheck = now;
+    const previousNeighbors = new Set(this.neighbors);
+    this.neighbors.clear();
+
+    // Find current neighbors
+    for (const box of allBoxes) {
+      if (this.isTouching(box)) {
+        this.neighbors.add(box);
+      }
+    }
+
+    // Notify about new neighbors
+    this.neighbors.forEach((neighbor) => {
+      if (!previousNeighbors.has(neighbor)) {
+        this.onNeighborAdded(neighbor);
+      }
+    });
+
+    // Notify about lost neighbors
+    previousNeighbors.forEach((previousNeighbor) => {
+      if (!this.neighbors.has(previousNeighbor)) {
+        this.onNeighborRemoved(previousNeighbor);
+      }
+    });
+  }
+
+  /**
+   * Check for intersections with boxes on different z-levels
+   */
+  public updateIntersections(allBoxes: SelectionBox[]): void {
+    const previous = new Set(this.intersections);
+    this.intersections.clear();
+
+    const myBounds = this.getBounds();
+    for (const box of allBoxes) {
+      if (box === this) continue;
+      if (box.getZLevel && box.getZLevel() === this.zLevel) continue; // only different z-levels
+      try {
+        const otherBounds = box.getBounds();
+        if (
+          Phaser.Geom.Intersects.RectangleToRectangle(myBounds, otherBounds)
+        ) {
+          this.intersections.add(box);
+        }
+      } catch (e) {
+        // skip boxes that don't implement getBounds
+      }
+    }
+
+    // If intersections changed, update visuals
+    if (this.intersections.size !== previous.size) {
+      this.updateTabWithNetworkInfo();
+    }
+  }
+
+  /**
+   * Called when a new neighbor is detected
+   */
+  private onNeighborAdded(neighbor: SelectionBox): void {
+    console.log(
+      `Box ${this.localContext.id} gained neighbor ${neighbor.localContext.id}`,
+    );
+    // Share our shareable data with the new neighbor
+    this.shareDataWithNeighbor(neighbor);
+    // Update visual indicator
+    this.updateTabWithNetworkInfo();
+  }
+
+  /**
+   * Called when a neighbor is no longer touching
+   */
+  private onNeighborRemoved(neighbor: SelectionBox): void {
+    console.log(
+      `Box ${this.localContext.id} lost neighbor ${neighbor.localContext.id}`,
+    );
+    // Update visual indicator
+    this.updateTabWithNetworkInfo();
+  }
+
+  /**
+   * Share shareable data with a specific neighbor
+   */
+  private shareDataWithNeighbor(neighbor: SelectionBox): void {
+    const shareableData = this.getShareableData();
+    shareableData.forEach((data, key) => {
+      neighbor.receiveSharedData(key, data);
+    });
+  }
+
+  /**
+   * Get current neighbors
+   */
+  public getNeighbors(): SelectionBox[] {
+    return Array.from(this.neighbors);
+  }
+
+  // STEP 4: Collaborative Context Merging - Data sharing and merging methods
+
+  /**
+   * Receive shared data from a neighbor - implements collaborative merging
+   */
+  public receiveSharedData(key: string, incomingData: BoxContextData): void {
+    const existingData = this.localContext.data.get(key);
+
+    if (!existingData) {
+      // We don't have this data, accept it if it's shareable
+      if (incomingData.canShare) {
+        this.localContext.data.set(key, {
+          ...incomingData,
+          // Don't change ownership when receiving shared data
+        });
+        this.localContext.version++;
+      }
+      return;
+    }
+
+    // We have this data - apply collaborative merging rules
+    if (existingData.owner === this.localContext.id) {
+      // We own this data - only update if incoming is newer from same owner
+      if (
+        incomingData.owner === existingData.owner &&
+        incomingData.version > existingData.version
+      ) {
+        this.localContext.data.set(key, incomingData);
+        this.localContext.version++;
+      }
+      // Otherwise, keep our version (we're the owner)
+    } else {
+      // We don't own this data - accept updates from the rightful owner
+      if (
+        incomingData.owner === existingData.owner &&
+        incomingData.version > existingData.version
+      ) {
+        this.localContext.data.set(key, incomingData);
+        this.localContext.version++;
+      }
+    }
+  }
+
+  /**
+   * Broadcast shareable data to all current neighbors
+   */
+  public broadcastToNeighbors(): void {
+    const shareableData = this.getShareableData();
+    this.neighbors.forEach((neighbor) => {
+      shareableData.forEach((data, key) => {
+        neighbor.receiveSharedData(key, data);
+      });
+    });
+  }
+
+  /**
+   * Request specific data from neighbors
+   */
+  public requestDataFromNeighbors(key: string): BoxContextData | null {
+    for (const neighbor of this.neighbors) {
+      const data = neighbor.localContext.data.get(key);
+      if (data && data.canShare) {
+        // Receive this data through normal merging process
+        this.receiveSharedData(key, data);
+        return data;
+      }
+    }
+    return null;
   }
 
   getZLevel(): number {
@@ -113,67 +415,67 @@ export class SelectionBox {
     // Pick color based on zLevel
     const color = this.getColorForZLevel(this.zLevel);
 
-    // Draw filled region
+    // Visual margin to include the tab/indicator above the box (in pixels)
+    const tabVisualMargin = 14;
+
+    // Pixel coordinates for drawing (expand upward by tabVisualMargin)
+    const pixelLeft = startX * 16;
+    const pixelTop = startY * 16 - tabVisualMargin;
+    const pixelRight = endX * 16 + 16;
+    const pixelBottom = endY * 16 + 16;
+    const pixelWidth = pixelRight - pixelLeft;
+    const pixelHeight = pixelBottom - pixelTop;
+
+    // Draw filled region (including margin for the tab)
     this.graphics.fillStyle(color, 0.3);
-    this.graphics.fillRect(
-      startX * 16,
-      startY * 16,
-      (endX - startX + 1) * 16,
-      (endY - startY + 1) * 16,
-    );
+    this.graphics.fillRect(pixelLeft, pixelTop, pixelWidth, pixelHeight);
 
     // Draw dashed border
     this.graphics.lineStyle(2, color, 1);
     this.graphics.beginPath();
 
-    const width = endX - startX + 1;
-    const height = endY - startY + 1;
+    // width/height (in tiles) were previously used for dashed border math; replaced by pixelWidth/pixelHeight
     const dashLength = 8;
     const gapLength = 4;
 
     if (this.isFinalized) {
-      // Solid rectangle border for finalized boxes
-      this.graphics.strokeRect(
-        startX * 16,
-        startY * 16,
-        (endX - startX + 1) * 16,
-        (endY - startY + 1) * 16,
-      );
+      // Solid rectangle border for finalized boxes (include margin)
+      this.graphics.strokeRect(pixelLeft, pixelTop, pixelWidth, pixelHeight);
     } else {
       // Dashed border for temporary boxes
-      // Top border
-      for (let i = 0; i < width * 16; i += dashLength + gapLength) {
-        this.graphics.moveTo(startX * 16 + i, startY * 16);
+      // Top border (use pixelTop so the dashed border includes the tab area)
+      for (let i = 0; i < pixelWidth; i += dashLength + gapLength) {
+        this.graphics.moveTo(pixelLeft + i, pixelTop);
         this.graphics.lineTo(
-          Math.min(startX * 16 + i + dashLength, endX * 16 + 16),
-          startY * 16,
+          Math.min(pixelLeft + i + dashLength, pixelRight),
+          pixelTop,
         );
       }
 
       // Bottom border
-      for (let i = 0; i < width * 16; i += dashLength + gapLength) {
-        this.graphics.moveTo(startX * 16 + i, endY * 16 + 16);
+      for (let i = 0; i < pixelWidth; i += dashLength + gapLength) {
+        this.graphics.moveTo(pixelLeft + i, pixelBottom);
         this.graphics.lineTo(
-          Math.min(startX * 16 + i + dashLength, endX * 16 + 16),
-          endY * 16 + 16,
+          Math.min(pixelLeft + i + dashLength, pixelRight),
+          pixelBottom,
         );
       }
 
       // Left border
-      for (let i = 0; i < height * 16; i += dashLength + gapLength) {
-        this.graphics.moveTo(startX * 16, startY * 16 + i);
+      for (let i = 0; i < pixelHeight; i += dashLength + gapLength) {
+        this.graphics.moveTo(pixelLeft, pixelTop + i);
         this.graphics.lineTo(
-          startX * 16,
-          Math.min(startY * 16 + i + dashLength, endY * 16 + 16),
+          pixelLeft,
+          Math.min(pixelTop + i + dashLength, pixelBottom),
         );
       }
 
       // Right border
-      for (let i = 0; i < height * 16; i += dashLength + gapLength) {
-        this.graphics.moveTo(endX * 16 + 16, startY * 16 + i);
+      for (let i = 0; i < pixelHeight; i += dashLength + gapLength) {
+        this.graphics.moveTo(pixelRight, pixelTop + i);
         this.graphics.lineTo(
-          endX * 16 + 16,
-          Math.min(startY * 16 + i + dashLength, endY * 16 + 16),
+          pixelRight,
+          Math.min(pixelTop + i + dashLength, pixelBottom),
         );
       }
 
@@ -328,6 +630,28 @@ export class SelectionBox {
             }
           }
           if (!intersects) {
+            // compute integer delta relative to initial start
+            const deltaX = Math.floor(
+              candidateStart.x - (this._dragInitialStart?.x ?? 0),
+            );
+            const deltaY = Math.floor(
+              candidateStart.y - (this._dragInitialStart?.y ?? 0),
+            );
+            // update placedTiles positions relative to the original snapshot
+            try {
+              if (
+                this._dragOriginalPlacedTiles &&
+                this._dragOriginalPlacedTiles.length > 0
+              ) {
+                this.placedTiles = this._dragOriginalPlacedTiles.map((p) => ({
+                  tileIndex: p.tileIndex,
+                  x: p.x + deltaX,
+                  y: p.y + deltaY,
+                  layerName: p.layerName,
+                }));
+              }
+            } catch (e) {}
+
             this.start = candidateStart;
             this.end = candidateEnd;
             this.redraw();
@@ -351,6 +675,127 @@ export class SelectionBox {
         try {
           if (this._pointerUpHandler)
             this.scene.input.off("pointerup", this._pointerUpHandler);
+        } catch (e) {}
+        // If we had a snapshot of placed tiles, commit a move of those tiles on the map
+        try {
+          const orig = this._dragOriginalPlacedTiles;
+          if (orig && orig.length > 0) {
+            const rg = (this.scene as any).regenerator as any;
+            // compute delta from initial drag start if needed
+            const deltaX = Math.floor(
+              this.start.x - (this._dragInitialStart?.x ?? 0),
+            );
+            const deltaY = Math.floor(
+              this.start.y - (this._dragInitialStart?.y ?? 0),
+            );
+
+            const movements: Array<any> = [];
+            if (this.placedTiles && this.placedTiles.length === orig.length) {
+              for (let i = 0; i < orig.length; i++) {
+                const o = orig[i];
+                const n = this.placedTiles[i];
+                if (!n) continue;
+                // only add movement if different
+                if (o.x === n.x && o.y === n.y) continue;
+                movements.push({
+                  type: "tile",
+                  layerName: o.layerName,
+                  from: { x: o.x, y: o.y },
+                  to: { x: n.x, y: n.y },
+                  index: o.tileIndex,
+                });
+              }
+            } else {
+              // fallback: apply delta to each original
+              for (const o of orig) {
+                const toX = o.x + deltaX;
+                const toY = o.y + deltaY;
+                if (o.x === toX && o.y === toY) continue;
+                movements.push({
+                  type: "tile",
+                  layerName: o.layerName,
+                  from: { x: o.x, y: o.y },
+                  to: { x: toX, y: toY },
+                  index: o.tileIndex,
+                });
+              }
+            }
+
+            if (
+              movements.length > 0 &&
+              rg &&
+              typeof rg.moveObjects === "function"
+            ) {
+              try {
+                rg.moveObjects(movements);
+              } catch (e) {
+                // ignore move errors but log for debugging
+                // eslint-disable-next-line no-console
+                console.error("SelectionBox: moveObjects failed", e);
+              }
+            }
+          }
+        } catch (e) {
+          // swallow
+        }
+        // Also handle any placed enemies that were moved with the box
+        try {
+          const origE = this._dragOriginalPlacedEnemies;
+          if (origE && origE.length > 0) {
+            const rg = (this.scene as any).regenerator as any;
+            const deltaX = Math.floor(
+              this.start.x - (this._dragInitialStart?.x ?? 0),
+            );
+            const deltaY = Math.floor(
+              this.start.y - (this._dragInitialStart?.y ?? 0),
+            );
+            const eMoves: Array<any> = [];
+            if (
+              this.placedEnemies &&
+              this.placedEnemies.length === origE.length
+            ) {
+              for (let i = 0; i < origE.length; i++) {
+                const o = origE[i];
+                const n = this.placedEnemies[i];
+                if (!n) continue;
+                if (o.x === n.x && o.y === n.y) continue;
+                eMoves.push({
+                  type: "enemy",
+                  from: { x: o.x, y: o.y },
+                  to: { x: n.x, y: n.y },
+                });
+              }
+            } else {
+              for (const o of origE) {
+                const toX = o.x + deltaX;
+                const toY = o.y + deltaY;
+                if (o.x === toX && o.y === toY) continue;
+                eMoves.push({
+                  type: "enemy",
+                  from: { x: o.x, y: o.y },
+                  to: { x: toX, y: toY },
+                });
+              }
+            }
+            if (
+              eMoves.length > 0 &&
+              rg &&
+              typeof rg.moveObjects === "function"
+            ) {
+              try {
+                rg.moveObjects(eMoves);
+              } catch (er) {
+                // eslint-disable-next-line no-console
+                console.error("SelectionBox: moveObjects (enemies) failed", er);
+              }
+            }
+          }
+        } catch (e) {
+          // swallow
+        }
+        // clear drag snapshot
+        try {
+          this._dragOriginalPlacedTiles = undefined;
         } catch (e) {}
 
         // Only commit if we actually started a drag snapshot - Drag and Drop
@@ -380,6 +825,23 @@ export class SelectionBox {
           // prepare drag
           this._dragInitialStart = this.start.clone();
           this._dragInitialEnd = this.end.clone();
+          // snapshot placed tiles so we can update their coordinates while dragging
+          try {
+            this._dragOriginalPlacedTiles = this.placedTiles.map((p) => ({
+              tileIndex: p.tileIndex,
+              x: p.x,
+              y: p.y,
+              layerName: p.layerName,
+            }));
+            this._dragOriginalPlacedEnemies = this.placedEnemies.map((e) => ({
+              enemyType: e.enemyType,
+              x: e.x,
+              y: e.y,
+            }));
+          } catch (err) {
+            this._dragOriginalPlacedTiles = undefined;
+            this._dragOriginalPlacedEnemies = undefined;
+          }
           const cam =
             this.scene.cameras && this.scene.cameras.main
               ? this.scene.cameras.main
@@ -560,6 +1022,294 @@ export class SelectionBox {
     // nothing else needed for now
   }
 
+  // Collaborative Context Merging - Helper methods for easy usage
+
+  /**
+   * Get a summary of all data available in our network
+   * @returns Object with our data and neighbors' shareable data
+   */
+  public getNetworkDataSummary(): {
+    own: string[];
+    neighborsShareable: string[];
+  } {
+    const own = Array.from(this.localContext.data.keys());
+    const neighborsShareable = new Set<string>();
+
+    this.neighbors.forEach((neighbor) => {
+      neighbor.localContext.data.forEach((data, key) => {
+        if (data.canShare) {
+          neighborsShareable.add(key);
+        }
+      });
+    });
+
+    return {
+      own,
+      neighborsShareable: Array.from(neighborsShareable),
+    };
+  }
+
+  /**
+   * Get basic info about this box for debugging
+   */
+  public getDebugInfo(): any {
+    return {
+      id: this.localContext.id,
+      zLevel: this.zLevel,
+      bounds: this.getBounds(),
+      neighbors: this.neighbors.size,
+      dataKeys: Array.from(this.localContext.data.keys()),
+      version: this.localContext.version,
+    };
+  }
+
+  // Collaborative Context Merging - Demo/Test methods
+
+  /**
+   * Demo method: Set some test data and share it with neighbors
+   * This shows how the collaborative system works
+   */
+  public demoCollaborativeSharing(): void {
+    console.log(`Box ${this.localContext.id} starting demo...`);
+
+    // Set some shareable data
+    this.shareData("demo_message", `Hello from ${this.localContext.id}!`, true);
+    this.shareData("timestamp", Date.now(), true);
+    this.shareData("box_color", this.getColorForZLevel(this.zLevel), true);
+
+    // Set some private data (not shareable)
+    this.setContextData("private_note", "This is private data", false);
+
+    console.log(
+      `Box ${this.localContext.id} shared data with ${this.neighbors.size} neighbors`,
+    );
+  }
+
+  /**
+   * Demo method: Log all available data in the network
+   */
+  public demoLogNetworkData(): void {
+    console.log(`=== Network Data for Box ${this.localContext.id} ===`);
+    console.log("My data:", Array.from(this.localContext.data.entries()));
+    console.log("Network summary:", this.getNetworkDataSummary());
+    console.log("Debug info:", this.getDebugInfo());
+
+    this.neighbors.forEach((neighbor) => {
+      console.log(
+        `Neighbor ${neighbor.localContext.id}:`,
+        neighbor.getDebugInfo(),
+      );
+    });
+  }
+
+  // Collaborative Context Merging - Chat system integration
+
+  /**
+   * Get collaborative context information for the language model
+   * This provides rich context about the box and its neighbors
+   */
+  public getCollaborativeContextForChat(): string {
+    const contextLines: string[] = [];
+
+    // Basic box info
+    contextLines.push(`=== Box Context ===`);
+    contextLines.push(`Box ID: ${this.localContext.id}`);
+    contextLines.push(`Z-Level: ${this.zLevel}`);
+    contextLines.push(
+      `Position: (${this.start.x}, ${this.start.y}) to (${this.end.x}, ${this.end.y})`,
+    );
+    contextLines.push(`Neighbors: ${this.neighbors.size} connected boxes`);
+
+    // Own data
+    if (this.localContext.data.size > 0) {
+      contextLines.push(`\n=== My Data ===`);
+      this.localContext.data.forEach((data, key) => {
+        contextLines.push(
+          `${key}: ${JSON.stringify(data.value)} (${data.canShare ? "shareable" : "private"})`,
+        );
+      });
+    }
+
+    // Network data summary
+    const networkSummary = this.getNetworkDataSummary();
+    if (networkSummary.neighborsShareable.length > 0) {
+      contextLines.push(`\n=== Available from Neighbors ===`);
+      networkSummary.neighborsShareable.forEach((key) => {
+        const value = this.findData(key);
+        if (value !== null) {
+          contextLines.push(`${key}: ${JSON.stringify(value)} (from neighbor)`);
+        }
+      });
+    }
+
+    // Neighbor details
+    if (this.neighbors.size > 0) {
+      contextLines.push(`\n=== Neighbor Details ===`);
+      this.neighbors.forEach((neighbor) => {
+        const neighborInfo = neighbor.getDebugInfo();
+        contextLines.push(
+          `Neighbor ${neighborInfo.id}: Z${neighborInfo.zLevel}, ${neighborInfo.dataKeys.length} data items`,
+        );
+      });
+    }
+
+    return contextLines.join("\n");
+  }
+
+  /**
+   * Add a chat message and optionally share it with neighbors
+   * @param msg - The message to add
+   * @param shareWithNeighbors - Whether to share this message with touching boxes
+   */
+  public addCollaborativeChatMessage(
+    msg: any,
+    shareWithNeighbors: boolean = false,
+  ): void {
+    // Add to our own chat history
+    this.localContext.chatHistory.push(msg);
+
+    // Optionally share with neighbors
+    if (shareWithNeighbors && typeof msg.content === "string") {
+      this.shareData(
+        "last_chat_message",
+        {
+          content: msg.content,
+          timestamp: Date.now(),
+          from: this.localContext.id,
+        },
+        true,
+      );
+    }
+  }
+
+  /**
+   * Get shared chat messages from neighbors
+   */
+  public getSharedChatMessages(): any[] {
+    const sharedMessages: any[] = [];
+
+    this.neighbors.forEach((neighbor) => {
+      const sharedMsg = neighbor.getContextData("last_chat_message");
+      if (sharedMsg) {
+        sharedMessages.push({
+          ...sharedMsg,
+          fromNeighbor: neighbor.localContext.id,
+        });
+      }
+    });
+
+    return sharedMessages.sort((a, b) => a.timestamp - b.timestamp);
+  }
+
+  // Testing and Verification Methods
+
+  /**
+   * Quick test: Set some test data and verify neighbors can see it
+   */
+  public testCollaborativeSharing(): void {
+    console.log(
+      `🧪 Testing collaborative sharing for Box ${this.localContext.id}`,
+    );
+
+    // Set some test data
+    this.shareData(
+      "test_message",
+      `Hello from Box ${this.localContext.id}!`,
+      true,
+    );
+    this.shareData("test_number", Math.floor(Math.random() * 100), true);
+    this.shareData("test_timestamp", new Date().toISOString(), true);
+
+    console.log(`📤 Box ${this.localContext.id} shared 3 test items`);
+    console.log(`👥 Connected to ${this.neighbors.size} neighbors`);
+
+    // Log what we can see from neighbors
+    setTimeout(() => {
+      const summary = this.getNetworkDataSummary();
+      console.log(
+        `📥 Box ${this.localContext.id} can see from neighbors:`,
+        summary.neighborsShareable,
+      );
+    }, 100);
+  }
+
+  /**
+   * Verify the collaborative context is working for chat
+   */
+  public testChatContext(): string {
+    console.log(`💬 Testing chat context for Box ${this.localContext.id}`);
+    const context = this.getCollaborativeContextForChat();
+    console.log("Generated context:", context);
+    return context;
+  }
+
+  /**
+   * Visual indicator: Change tab color based on neighbor count
+   */
+  public updateTabWithNetworkInfo(): void {
+    if (!this.tabText) return;
+
+    const neighborCount = this.neighbors.size;
+    const dataCount = this.localContext.data.size;
+    const intersectionCount = this.intersections.size;
+
+    // Update tab text to show network info, but omit zero-valued parts
+    const parts: string[] = [];
+    if (neighborCount > 0) parts.push(`${neighborCount}n`);
+    if (dataCount > 0) parts.push(`${dataCount}d`);
+    if (intersectionCount > 0) parts.push(`${intersectionCount}z`);
+    const text = parts.length > 0 ? `Box (${parts.join(", ")})` : `Box`;
+    this.tabText.setText(text);
+
+    // Make the tab wider when visual indicators are added: add 20px per indicator
+    if (this.tabBg && this.tabText) {
+      const baseWidth = 48;
+      const padding = 8;
+      const indicatorsExtra = parts.length * 3;
+      const measured = Math.ceil(this.tabText.width || 0) + padding * 2;
+      const newWidth = Math.max(baseWidth, measured + indicatorsExtra);
+
+      try {
+        // Update rectangle display size (some Phaser builds support setDisplaySize)
+        if (typeof (this.tabBg as any).setDisplaySize === "function") {
+          (this.tabBg as any).setDisplaySize(
+            newWidth,
+            (this.tabBg as any).height || 14,
+          );
+        } else {
+          // Fallback: directly set width property
+          (this.tabBg as any).width = newWidth;
+        }
+      } catch (e) {
+        // ignore if resizing not supported
+      }
+
+      // Update container size and text padding
+      this.tabContainer?.setSize(newWidth, (this.tabBg as any).height || 14);
+      // Keep text positioned with left padding
+      try {
+        this.tabText.x = padding;
+      } catch (e) {}
+    }
+
+    // Change color based on connectivity
+    if (this.tabBg) {
+      if (intersectionCount > 0) {
+        // Intersection present - orange stroke to warn
+        this.tabBg.setFillStyle(this.isActive ? 0xffe0cc : 0xfff0e6);
+        this.tabBg.setStrokeStyle(2, 0xffa500);
+      } else if (neighborCount > 0) {
+        // Connected - use cyan to indicate network activity
+        this.tabBg.setFillStyle(this.isActive ? 0x00ff88 : 0x00aaff);
+      } else {
+        // Not connected - use original colors
+        this.tabBg.setFillStyle(
+          this.isActive ? 0x127803 : this.isFinalized ? 0x2b2b2b : 0x2b6bff,
+        );
+      }
+    }
+  }
+
   // Chat history management for this selection box
   addChatMessage(msg: any) {
     this.localContext.chatHistory.push(msg);
@@ -592,6 +1342,16 @@ export class SelectionBox {
     console.log("Added placed tile:", { tileIndex, x, y, layerName });
   }
 
+  // Register an enemy placed within this selection box (tile coords)
+  public addPlacedEnemy(enemyType: string, x: number, y: number) {
+    this.placedEnemies.push({ enemyType, x, y });
+    console.log("Added placed enemy:", { enemyType, x, y });
+  }
+
+  public getPlacedEnemies() {
+    return this.placedEnemies;
+  }
+
   public getPlacedTiles() {
     return this.placedTiles;
   }
@@ -601,6 +1361,53 @@ export class SelectionBox {
     this.placedTiles.forEach((tile, index) => {
       console.log(`${index + 1}: ${JSON.stringify(tile)}`);
     });
+  }
+
+  public shareData(key: string, value: any, canShare: boolean = true): void {
+    this.setContextData(key, value, canShare);
+    this.broadcastToNeighbors();
+  }
+
+  /**
+   * Get data, checking neighbors if we don't have it locally
+   * @param key - The data key to look for
+   * @returns The data value or null if not found
+   */
+  public findData(key: string): any {
+    // Check our own data first
+    const localData = this.getContextData(key);
+    if (localData !== null) {
+      return localData;
+    }
+
+    // Request from neighbors if we don't have it
+    const neighborData = this.requestDataFromNeighbors(key);
+    return neighborData ? neighborData.value : null;
+  }
+
+  /**
+   * Check if any box in our network (us or neighbors) has specific data
+   * @param key - The data key to check for
+   * @returns true if any connected box has this data
+   */
+  public networkHasData(key: string): boolean {
+    // Check ourselves first
+    if (this.localContext.data.has(key)) {
+      return true;
+    }
+
+    // Check neighbors
+    for (const neighbor of this.neighbors) {
+      if (neighbor.localContext.data.has(key)) {
+        return true;
+      }
+    }
+
+    return false;
+  }
+
+  public getActive(): boolean {
+    return this.isActive;
   }
 
   //Drag and Drop support - Jason Cho
